Implement highlighting based on byte offsets from payloads
This is a large, sweeping change to introduce support for highlighting
based on externally-located UTF8 files.

The basic approach is to replace the "raw" OCR input document with a
simplified format that encodes a word alongside its byte offsets in the
original UTF8 file.

During indexing, the offset is stored in the term payload. At
highlighting time, we use a custom highlighting strategy (based on
the UnifiedHighlighter stack) to use these byte offsets to perform
highlighting.

This also includes a byte offset parser for hOCR (i.e. a method to
generate source documents with byte offsets from hOCR documents),
however this cannot yet be used for highlighting, since we're still
missing a OcrPassageFormatter and BreakIterator implementations for it.
Support for ALTOL should also be added in the future.

Note that the plugin still supports highlighting based on UTF-16
codepoints stored in the index or on disk.