Fix calculation of passages with matches spanning multiple lines
Previously a passage would be determined by the context of the first
match it contained (e.g. previous two lines + line with match + next two
lines). This caused breakage in cases where another match occurred on
the last line of the previous match's context.

We now expand the passage to the end of the last match's context.

Additionally, the end of a context is now determined based on the end of
the match, not the start. This will add one or multiple additional
context lines for matches that span multiple lines.