Allow loading of files at indexing time on Solr-side
This is the beginning of a changeset that should drastically simplify
the workings of the plugin. The basic idea is this:
Since we need to access the OCR files on disk anyway at query time, why
not load them from there for indexing purposes as well?
The user can now just pass a file pointer in the OCR field at indexing
time, the plugin will use the pointer to load the OCR content for
analysis and from there the plugin works just as before.
The advantages are numerous:

- Users don't have to re-encode their files to ASCII any more, since we
  can now determine the byte offsets ourselves during indexing
- Users don't have to push ginormous piles of XML over the pipe anymore,
  only a very lightweight file pointer string
- Only two ways of using the plugin now: OCR stored in Solr or OCR
  stored on disk. No more questions of file encoding
- We can probably throw away >50% of the code, which means less stuff to
  maintain and less surface for bugs to appear.