Load source pointers from stored field content
This makes it now mandatory for OCR fields to be stored, even though
the actual content comes from an external source.
The advantage of this approach that we can completely ditch the
`ExternalFieldLoader` interface with its rather complex resolving
mechanism, since the paths are now directly stored in the index.

One less thing to maintain, one less thing to cause bugs.