Add `hl.ocr.maxPassages` parameter to improve performance
Solr will by default calculate all passages from a document and then
score and rank them as if they were separate documents. For queries that
result in many matches in a document this can get expensive, since we
potentially need to scan the input files for more than 100 passages,
only to finally use a single one.

So this commit introduces a new parameter to limit the number of
snippets the plugin has to generate. By default the value is 100, but if
the user's `hl.snippets` value is higher, it will be used instead.