Update to Solr 8.4, while maintaining compatbility with older versions.
This requires another round of reflection-based legacy API adapters.
We really don't want to mainting multiple parallel branches for various
legacy versions of Solr, so this is the way for now.

This is probably going to get out of hand after a few upstream release
cycles, so we'll likely have to reduce to one or two supported legacy
version at some point.  Candidates could be pre-8.1 and post-8.4.