Add new `pages` key to snippet to provide page dimensions.
This can be useful for instances where the coordinates in the OCR do not match the
actual page image that is being highlighted (e.g. because it is scaled down for
delivery). With knowledge of the page dimensions of the OCR page, users can then
calculate the required conversions for themselves.

The runtime impact of this change should be minimal, since we're not reading in any
additional parts of the input documents, we need to parse the page elements anyway
to obtain the page identifier.