Add timeAllowed parameter to limit highlighting time
The timeout will be checked during the most taxing part of the
highlighting process, i.e. while reading the input OCR files from disk.
If the timeout is encountered, highlighting for the current and any
further documents is cancelled and the highlighting terminated. Any
snippets for documents that were handled before the timeout will be
included in the response. The `responseHeader` will indicate the
partiality of the results with the `partialResults` flag.

**Note** that to get the same behavior in a distributed search, you have
to ensure two things:

- `shards.tolerant` is set to `true`, otherwise a timeout in one shard
  will fail the complete query
- `sort` contains at least one actual document field (i.e. not just
  `score`), otherwise your query will be successful, but you will see
  no actual result documents or highlights, despite them having
  completed successfully.