Dirty speed hack: Ignore Unicode during forward/backward seeking
This hack modifies the `charAt` method on `FileBytesCharIterator` to
just return `?` if the byte at the position is non-ASCII. This gives us
a ~25% performance boost over the old proper way of handling UTF-8
during iteration, without having much of an effect, since this method is
only called to determine the breaks. Since all of the supported OCR
formats use plain ASCII for their element and attribute names, this
should be safe. `subSequence`, wich is used to actually get the content,
still includes proper UTF-8 handling, but this is off the hot path and
thus not critical for performance.