Fix highlighting for phrase matches crossing columns
This was a tricky one. It can happen that a phrase match crosses column
boundaries, i.e. one part is at the bottom of the left column and
another part is at the top of the right column.
In these cases we now return a snippet with two regions. The highlights
in each span now have a new `regionIdx` field that references the
snippet region they belong to. This replaces the `pageId` field, this
value can now be obtained from the referenced region.