Allow peeking back into the past with PeekingReader.
This allows us to correctly determine the offset of content attributes
when parsing ALTO, without having to make fragile assumptions about the
input formatting, i.e. if the ALTO is valid XML, it should parse fine,
no matter how freaky the formatting is (well, Ok, if you use more than
16k chars of whitespace inside of a <String /> it's going to break...)

This change made it neccessary for OCR CharFilters to require a
`PeekingReader` as the input reader, but the `OcrCharFilterFactory`
takes care of that.