Improve handling of HTML entities used in hOCR.
Previously we would resolve e.g. `&shy;` with a callback. This was a)
not really performant since Woodstax would cause a `stat()` syscall for
every unresolved entity under the hood, causing a huge perfomance hit
and b) missed a few entities that are used in hOCR.

We now use Woodstax' support for custom entities to pre-register these
entities during parser construction.