Fix implicit default highlighting fields (implements #152).
This was an ugly limitation of the previous version, due to which users
had always explicitely pass a list of 'regular' highlighting fields via
`hl.fl`. A refactoring of the `OcrHighlightComponent` allows us to
remedy this.

The fix is a bit hacky, but it works reliably: When we detect that
highlighting is enabled, but no `hl.fl` parameter was passed (i.e.
the highlighter will automatically highlight all stored fields), we
create a custom explicit `hl.fl` parameter at runtime, that includes all
of the default highlight fields, except for those fields contained in
the `hl.ocr.fl` field set.