Fix oversized highlights in some cases (fixes #213) (#216)
This was spotted by @MichaelHoppe (thanks!!!): If a matched word is
followed by another word that only consists of punctuation, the
punctuation-only word will be part of the match, despite not
contributing anything to its content (i.e. the end of the highlight span
would be before the first character of the word). This would lead to
highlights with the  correct textual contents, but coordinates that were
off. This is now fixed, we only add words to a highlight span if they
don't start with the end-of-hl marker.