sanitization: Make 'end of tag' detection more robust.
Previously we would look for the first space (0x20) character within a
tag to determine where the element name ended. This is not robust, since
a given element may be split across multiple lines or include tab
characters instead of a simple whitesapce. This is why we now check for
those as well, hopefully making the XML sanitizer more robust.
Thanks to @schmika for spotting the bug!