Add support for Solr 9.
From this change on, we will have to release two separate JAR files, one
for Solr 9 and later, and one for Solr 8 and earlier. The reason for
this is that with Lucene 9, the package path for `CharFilterFactory` has
changed, which is something we can't work around with Reflection magic
like we did so far. Additionally, `CloseHook` became an interface in
Solr, instead of an abstract class, with the same problem.

So we now provide a script in `util/patch_solr78_bytecode.py` to create
a Solr7/8 compatible JAR from one built against Solr 9.

We did not use ByteBuddy or ASM, since building multiple JARs from a
single POM is a ridiculous pain, so this little Python script to do
the neccessary bytecode manipulation manually seemed easier and simpler
to maintain.