Change namespace to `com.github.dbmdz` and `solrocr`.
- All internally used classes are moved from `de.digitalcollections` to
  `com.github.dbmdz`
- All user-visible classes are moved to the `solrocr` package namespace
  for less verbose configuration entry.

The `solrocr` namespace goes against Java best practices, but since this
plugin is not intended to be used as a library in other applications,
with only users editing the Solr configuration as the target audience,
this seems like a good reason to simplify the namespace overhead as much
as possible.

The rest of the documentation will be updated in a separate changeset.