Fix incomplete forward passes in HocrClassBreakLocator (fixes #288)
When locating breaks based on hOCR classes, we read the input in blocks
of 64k chars. Previously, due to an error in the looping logic, we would
not actually read any blocks besides the first one  and set the limit to
the end of the first block. This should be fixed now, more blocks are
read as needed.