Optimize read-buffer sizes during break locating.
If I/O is not an issue (e.g. because the data completely resides in the
page cache, then suddenly our generously sized 64KiB String buffer
becomes a problem due to the amount of memory copying that goes on when
constructing a String in Java (huge drawback of immutable strings!).
To optimize this, the buffer size was made dynamic, based on the block
type that is looked for, and users are offered customization options to
adapt the sizes to their data.