Fix race condition in OcrFieldHighlighter.
We were storing the number of matches in a non-thredsafe `HashMap`,
which led to a race condition when writing to it from multiple threads.
We now use a threadsafe `ConcurrentHashMap` to avoid this.

Additionally, we now use `CompleteableFuture`s for waiting on our
asynchronous highlighting threads instead of raw `Future`s.