Fix bad total snippets counts.
Turns out we were mixing up the counts between multiple documents by
keying according to the leaf-reader-relativ docId instead of the
index-absolute docId. This led to differing total counts between
different runs of one and the same query, which is how this was
discovered. Total counts should be accurate now.