Fix order of snippets when highlighting with scoring disabled.
The intention for this feature was to return the snippets in their order
of appearance in the highlighted document, with the earliest matches
coming first. Due to a mistake in the way this was implemented, the
order was correct, but users would get the *last* matches in the
document, not the *first* matches in the document.