feat: Allow parallel indexing of externally and internally stored OCR.
Previously we had to define two fields if we wanted to be flexible in
how we store the OCR, once for externally stored, and one for internally
stored. Upon reflection, this was an unneccessary limitation. It is now
possible to have a single field definition that allows for both cases.
In case the field value is not a valid source pointer, it will be
forwarded to the OCR parsing chain and be stored as-is in the document.
This allows for more flexible usage and more minimal schemas.