fix: Highlighting across multiple columns
A bug in the passage formatting code led to broken highlighting for
multi-column highlight spans, where the highlight span would only
contain a single highlight instead of two (one for each column). This
was due to an error in the way the column change heuristic was
calculated. This has now been fixed and highlight across multiple
columns is now working as intended.